Factor out field types and rename struct fields
