+++ p2panda-rs/src/schema/fields.rs
@@ -1,0 +1,87 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+use std::collections::BTreeMap;
+use std::fmt;
+
+/// CDDL types.
+#[derive(Clone, Debug, Copy)]
+#[allow(missing_docs)]
+pub enum Type {
+    Bool,
+    Int,
+    Float,
+    Tstr,
+    Relation,
+}
+
+/// CDDL schema type string formats.
+impl fmt::Display for Type {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        let cddl_type = match self {
+            Type::Bool => "bool",
+            Type::Int => "int",
+            Type::Float => "float",
+            Type::Tstr => "tstr",
+            Type::Relation => "tstr .regexp \"[0-9a-f]{68}\"",
+        };
+        write!(f, "{}", cddl_type)
+    }
+}
+
+/// CDDL field types.
+#[derive(Clone, Debug)]
+pub enum Field {
+    String(String),
+    Type(Type),
+    Struct(Group),
+}
+
+/// Format each different data type into a schema string.
+impl fmt::Display for Field {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            Field::String(s) => write!(f, "\"{}\"", s),
+            Field::Type(cddl_type) => write!(f, "{}", cddl_type),
+            Field::Struct(group) => write!(f, "{{ {} }}", group),
+        }
+    }
+}
+
+/// Struct for building and representing CDDL groups. CDDL uses groups to define reusable data
+/// structures they can be merged into schema or used in Vectors, Tables and Structs.
+#[derive(Clone, Debug)]
+pub struct Group {
+    #[allow(dead_code)] // Remove when schema module is used.
+    name: String,
+    fields: BTreeMap<String, Field>,
+}
+
+impl Group {
+    /// Create a new CDDL group.
+    pub fn new(name: String) -> Self {
+        Self {
+            name,
+            fields: BTreeMap::new(),
+        }
+    }
+
+    /// Add a field to the group.
+    pub fn add_field(&mut self, key: &str, field_type: Field) {
+        self.fields.insert(key.to_owned(), field_type);
+    }
+}
+
+impl fmt::Display for Group {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        let map = &self.fields;
+        write!(f, "( ")?;
+        for (count, value) in map.iter().enumerate() {
+            // For every element except the first, add a comma
+            if count != 0 {
+                write!(f, ", ")?;
+            }
+            write!(f, "{}: {}", value.0, value.1)?;
+        }
+        write!(f, " )")
+    }
+}
