+++ p2panda-rs/src/schema/schema.rs
@@ -11,6 +11,7 @@ use cddl::parser::Parser;
 use cddl::validate_cbor_from_slice;
 #[cfg(not(target_arch = "wasm32"))]
 use cddl::validator::cbor;
+use super::fields::{Field, Group, Type};
 
 use crate::hash::Hash;
 #[cfg(not(target_arch = "wasm32"))]
@@ -19,102 +20,11 @@ use crate::instance::{Instance, InstanceError};
 use crate::operation::{AsOperation, Operation, OperationFields, OperationValue};
 use crate::schema::SchemaError;
 
-/// CDDL types.
-#[derive(Clone, Debug, Copy)]
-#[allow(missing_docs)]
-pub enum Type {
-    Bool,
-    Int,
-    Float,
-    Tstr,
-    Relation,
-}
-
-/// CDDL schema type string formats.
-impl fmt::Display for Type {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let cddl_type = match self {
-            Type::Bool => "bool",
-            Type::Int => "int",
-            Type::Float => "float",
-            Type::Tstr => "tstr",
-            Type::Relation => "tstr .regexp \"[0-9a-f]{68}\"",
-        };
-        write!(f, "{}", cddl_type)
-    }
-}
-
-/// CDDL field types.
-#[derive(Clone, Debug)]
-pub enum Field {
-    String(String),
-    Type(Type),
-    Struct(Group),
-}
-
-/// Format each different data type into a schema string.
-impl fmt::Display for Field {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Field::String(s) => write!(f, "\"{}\"", s),
-            Field::Type(cddl_type) => write!(f, "{}", cddl_type),
-            Field::Struct(group) => write!(f, "{{ {} }}", group),
-        }
-    }
-}
-
-/// Struct for building and representing CDDL groups. CDDL uses groups to define reusable data
-/// structures they can be merged into schema or used in Vectors, Tables and Structs.
-#[derive(Clone, Debug)]
-pub struct Group {
-    #[allow(dead_code)] // Remove when schema module is used.
-    name: String,
-    fields: BTreeMap<String, Field>,
-}
-
-impl Group {
-    /// Create a new CDDL group.
-    pub fn new(name: String) -> Self {
-        Self {
-            name,
-            fields: BTreeMap::new(),
-        }
-    }
-
-    /// Add a field to the group.
-    pub fn add_field(&mut self, key: &str, field_type: Field) {
-        self.fields.insert(key.to_owned(), field_type);
-    }
-}
-
-impl fmt::Display for Group {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let map = &self.fields;
-        write!(f, "( ")?;
-        for (count, value) in map.iter().enumerate() {
-            // For every element except the first, add a comma
-            if count != 0 {
-                write!(f, ", ")?;
-            }
-            write!(f, "{}: {}", value.0, value.1)?;
-        }
-        write!(f, " )")
-    }
-}
-
 /// SchemaBuilder struct for programmatically creating CDDL schemas and validating OperationFields.
 #[derive(Clone, Debug)]
 pub struct SchemaBuilder {
     name: String,
     fields: BTreeMap<String, Field>,
-}
-
-/// Schema struct for creating CDDL schemas, validating OperationFields and creating operations
-/// following the defined schema.
-#[derive(Clone, Debug)]
-pub struct Schema {
-    schema_hash: Hash,
-    schema_string: String,
 }
 
 impl SchemaBuilder {
@@ -171,6 +81,14 @@ impl fmt::Display for SchemaBuilder {
     }
 }
 
+/// Schema struct for creating CDDL schemas, validating OperationFields and creating operations
+/// following the defined schema.
+#[derive(Clone, Debug)]
+pub struct Schema {
+    hash: Hash,
+    cddl: String,
+}
+
 impl Schema {
     /// Create a new Schema from a schema hash and schema CDDL string.
     pub fn new(schema_hash: &Hash, schema_str: &str) -> Result<Self, SchemaError> {
@@ -191,14 +109,14 @@ impl Schema {
         }?;
 
         Ok(Self {
-            schema_hash,
-            schema_string,
+            hash: schema_hash,
+            cddl: schema_string,
         })
     }
 
     /// Return the hash id of this schema.
     pub fn schema_hash(&self) -> Hash {
-        self.schema_hash.clone()
+        self.hash.clone()
     }
 
     /// Create a new CREATE operation validated against this schema.
@@ -278,7 +196,7 @@ impl Schema {
 
 impl fmt::Display for Schema {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.schema_string)
+        write!(f, "{}", self.cddl)
     }
 }
